+++ src/Modules/FoodSharePoint/FoodSharePointTransactions.php
@@ -9,8 +9,10 @@ use Foodsharing\Modules\Core\DBConstants\Bell\BellType;
 use Foodsharing\Modules\Core\DBConstants\Info\InfoType;
 use Foodsharing\Modules\Core\DBConstants\Uploads\UploadUsage;
 use Foodsharing\Modules\Uploads\UploadsGateway;
+use Foodsharing\Modules\Uploads\UploadsTransactions;
 use Foodsharing\Permissions\FoodSharePointPermissions;
 use Foodsharing\RestApi\Models\FoodSharePoint\AddFoodSharePointResponse;
+use Foodsharing\RestApi\Models\FoodSharePoint\FoodSharePointEditData;
 use Foodsharing\RestApi\Models\FoodSharePoint\FoodSharePointForCreation;
 use Foodsharing\RestApi\Models\Notifications\FoodSharePoint;
 use Foodsharing\Utility\EmailHelper;
@@ -24,6 +26,7 @@ class FoodSharePointTransactions
         private readonly FoodSharePointPermissions $foodSharePointPermissions,
         private readonly BellGateway $bellGateway,
         private readonly UploadsGateway $uploadsGateway,
+        private readonly UploadsTransactions $uploadsTransactions,
         private readonly EmailHelper $emailHelper,
         private readonly Sanitizer $sanitizer,
         private readonly TranslatorInterface $translator,
@@ -117,5 +120,27 @@ class FoodSharePointTransactions
         }
 
         return new AddFoodSharePointResponse($id, !$isProposal);
+    }
+
+    public function editFoodSharePoint(int $foodSharePointId, array $currentData, FoodSharePointEditData $newData): void
+    {
+        $this->foodSharePointGateway->updateFoodSharePoint($foodSharePointId, $newData);
+
+        /* If the picture of this food share point was changed, the usage type of the new one (if any) needs to be set
+         and the old picture needs to be deleted. */
+        $newPicture = $newData->picture ?? '';
+        if ($newPicture !== $currentData['picture']) {
+            if (!empty($currentData['picture'])) {
+                $oldUUID = substr($currentData['picture'], 13);
+                $this->uploadsTransactions->deleteUploadedFile($oldUUID);
+            }
+
+            if (!empty($newPicture)) {
+                $uuid = substr($newPicture, 13);
+                $this->uploadsGateway->setUsage([$uuid], UploadUsage::FOOD_SHARE_POINT_TITLE, $foodSharePointId);
+            }
+        }
+
+        $this->foodSharePointGateway->updateFSPManagers($currentData['id'], $newData->managerIds);
     }
 }