+++ tests/Acceptance/FoodsaverCest.php
@@ -70,7 +70,7 @@ class FoodsaverCest
         // Find an address in the search field
         $I->click('#change-address-button');
         $I->waitForText('Adresse auswählen');
-        $I->fillField('#searchinput', $address);
+        $I->fillField('#search-address-input', $address);
         $I->waitForElementVisible('#suggestions option');
         $I->click("//*[contains(text(), 'Teststraße 1')]");
         $I->click('Adresse übernehmen');